raspberrypi/requirements.txt
pyfcm==1.4.5
firebase_admin==2.16.0
dropbox==9.3.0
requests==2.21.0
beautifulsoup4==4.7.1
picamera==1.13
gevent_socketio==0.3.6
